Deduplicate URLs crafting
